requirements.txt
pandas==3.0.0
opencv-python==4.13.0.92
numpy==2.4.2
pytesseract==0.3.13
Pillow==12.1.0
torch
facenet-pytorch
transformers==5.1.0
matplotlib==3.10.8
rich
click
tqdm
scikit-learn
